Introduce automatic curried application.
Using the uncurried-always mode is difficult when calling compiler libraries, as those are uncurried.
This introduces automatic curried application in uncurried-only mode: when a curried function is used in application, the application is treated as curried application.
This means that even in uncurried-always mode, one can use compiler libraries in most cases. Exceptions are functions that take callbacks, such as `Array.map`.